@@ -18,71 +18,70 @@
     "eslint:fix": "npm run eslint -- --fix"
   },
   "dependencies": {
-    "@hookform/resolvers": "^2.8.1",
+    "@hookform/resolvers": "^2.8.3",
     "@hqoss/monads": "^0.5.0",
     "class-transformer": "^0.4.0",
     "class-validator": "^0.13.1",
     "clsx": "^1.1.1",
     "ensure-error": "3.X",
     "jwt-decode": "^3.1.2",
-    "ky": "^0.28.5",
+    "ky": "^0.28.7",
     "njct": "^8.1.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-hook-form": "^7.16.0",
-    "react-pending-resource": "^1.0.0",
-    "react-router-dom": "^5.3.0",
+    "react-hook-form": "^7.19.1",
+    "react-router-dom": "^6.0.1",
     "react-use-promise": "^0.4.1",
-    "recoil": "^0.4.1",
-    "use-promise": "^2.0.1"
+    "use-async-resource": "^2.2.2",
+    "recoil": "^0.5.2",
+    "suspend-react": "^0.0.8"
   },
   "devDependencies": {
     "@kyleshevlin/eslint-plugin": "^1.3.0",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.1.0",
+    "@testing-library/jest-dom": "^5.15.0",
+    "@testing-library/react": "^12.1.2",
     "@types/jest": "^27.0.2",
-    "@types/react": "^17.0.24",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-router-dom": "^5.3.0",
-    "@typescript-eslint/eslint-plugin": "^4.31.2",
-    "@typescript-eslint/parser": "^4.31.2",
+    "@types/react": "^17.0.34",
+    "@types/react-dom": "^17.0.11",
+    "@types/react-router-dom": "^5.3.2",
+    "@typescript-eslint/eslint-plugin": "^5.3.0",
+    "@typescript-eslint/parser": "^5.3.0",
     "@vitejs/plugin-react-refresh": "^1.3.6",
-    "autoprefixer": "^10.3.6",
-    "cssnano": "^5.0.8",
-    "cssnano-preset-default": "^5.1.4",
-    "eslint": "^7.32.0",
+    "autoprefixer": "^10.4.0",
+    "cssnano": "^5.0.10",
+    "cssnano-preset-default": "^5.1.6",
+    "eslint": "^8.2.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-import-resolver-node": "^0.3.6",
-    "eslint-plugin-etc": "^1.5.4",
-    "eslint-plugin-import": "^2.24.2",
+    "eslint-plugin-etc": "^2.0.1",
+    "eslint-plugin-import": "^2.25.2",
     "eslint-plugin-only-warn": "^1.0.3",
     "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-promise": "^5.1.0",
-    "eslint-plugin-react": "^7.26.0",
+    "eslint-plugin-promise": "^5.1.1",
+    "eslint-plugin-react": "^7.26.1",
     "eslint-plugin-react-hooks": "^4.2.0",
-    "eslint-plugin-regexp": "^1.3.1",
+    "eslint-plugin-regexp": "^1.5.0",
     "eslint-plugin-simple-import-sort": "^7.0.0",
     "eslint-plugin-sonarjs": "^0.10.0",
-    "eslint-plugin-sort-class-members": "^1.11.0",
-    "eslint-plugin-testing-library": "^4.12.4",
+    "eslint-plugin-sort-class-members": "^1.14.0",
+    "eslint-plugin-testing-library": "^5.0.0",
     "eslint-plugin-total-functions": "^4.10.1",
-    "eslint-plugin-unicorn": "^36.0.0",
+    "eslint-plugin-unicorn": "^38.0.1",
     "eslint-plugin-wix-editor": "^3.3.0",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^27.2.2",
+    "jest": "^27.3.1",
     "jest-mock-extended": "^2.0.4",
     "jest-transform-stub": "^2.0.0",
     "npm-run-all": "^4.1.5",
-    "postcss": "^8.3.8",
+    "postcss": "^8.3.11",
     "prettier": "^2.4.1",
     "simplytyped": "^3.3.0",
-    "tailwindcss": "^2.2.15",
-    "ts-jest": "^27.0.5",
+    "tailwindcss": "^2.2.19",
+    "ts-jest": "^27.0.7",
     "tsconfig-paths-jest-mapper": "^1.4.0",
-    "typescript": "^4.4.3",
-    "use-async-resource": "^2.2.1",
-    "vite": "^2.5.10",
-    "vite-tsconfig-paths": "^3.3.14",
+    "typescript": "^4.4.4",
+    "vite": "^2.6.13",
+    "vite-tsconfig-paths": "^3.3.17",
     "watchexec-bin": "^1.0.0"
   }
 }
