@@ -2,9 +2,9 @@ import { Interface } from '@libs/application';
 import { Article as SingleArticle } from '@libs/application/article';
 import { inject } from 'njct';
 import React, { useCallback, useEffect, useState } from 'react';
-import { RouteComponentProps } from 'react-router-dom';
 
 import { ArticleView } from './ArticleView';
+import { useParams } from 'react-router-dom';
 
 function useData(parameters: { slug: string }) {
   const { slug } = parameters;
@@ -26,8 +26,9 @@ function useData(parameters: { slug: string }) {
   return { article };
 }
 
-export function Article({ match }: RouteComponentProps<{ slug: string }>): JSX.Element {
-  const { article } = useData({ slug: match.params.slug });
+export function Article(): JSX.Element {
+  const { slug } = useParams();
+  const { article } = useData({ slug: slug! });
 
   return <ArticleView article={article} />;
 }
