@@ -1,6 +1,5 @@
 import { ArticleFeedHandler } from '@libs/application/article/queries';
 import React from 'react';
-import { RouteComponentProps } from 'react-router-dom';
 import usePromise from 'react-use-promise';
 
 import { HomeView } from './HomeView';
@@ -13,8 +12,7 @@ function useData() {
   return { articleList };
 }
 
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-export function Feed(_: RouteComponentProps): JSX.Element {
+export function Feed(): JSX.Element {
   const { articleList } = useData();
 
   return <HomeView articles={articleList?.articles} />;
