@@ -1,6 +1,5 @@
 import { ArticleListHandler } from '@libs/application/article/queries';
 import React from 'react';
-import { RouteComponentProps } from 'react-router-dom';
 import usePromise from 'react-use-promise';
 
 import { HomeView } from './HomeView';
@@ -13,7 +12,7 @@ function useData() {
   return { articleList };
 }
 
-export function Home(props: RouteComponentProps): JSX.Element {
+export function Home(): JSX.Element {
   const { articleList } = useData();
 
   return <HomeView articles={articleList?.articles} />;
